Cinema_BE/requirements.txt
flask
flask_sqlalchemy
flask_cors
pymysql
python-dotenv

